v1.9.0.1 - Fixed world gen issue with invalid config setting type

+++ Orbs/Config_Settings_Orbs_Sources.cs
@@ -73,7 +73,7 @@ namespace Orbs {
 
 		[Range( 1, 99 )]
 		[DefaultValue( 3 )]
-		public float MaxOrbsPerChestPerType { get; set; } = 3;
+		public int MaxOrbsPerChestPerType { get; set; } = 3;
 
 
 		//[DefaultValue(true)]
